+++ vue-spa/testing/string-utils.ts
@@ -1,5 +1,6 @@
 import { expect } from "vitest";
 import { type VueWrapper } from "@vue/test-utils";
+import { convertI18NInterpolationToRegex } from "../i18n-utils";
 
 interface ElementStringObject {
   placeholder?: string;
@@ -7,19 +8,6 @@ interface ElementStringObject {
   "data-testid"?: string;
   label?: string;
 }
-
-const regexCharacters = ["(", ")", ".", "*", "+", "?", "^", "$", "|", "/"];
-
-const convertI18NInterpolationToRegex = (str: string) => {
-  if (str.includes("{")) {
-    let escapedStr = str;
-    for (const char of regexCharacters) {
-      escapedStr = escapedStr.replace(char, `\\${char}`);
-    }
-    return new RegExp(escapedStr.replace(/\{(.*?)\}/g, ".*"));
-  }
-  return str;
-};
 
 // Store strings for Vue components in Record<string, ElementString>
 // Then you can v-bind them to the component, and also use them in tests for reliable element selection
